@@ -1,24 +1,21 @@
 {
-  "name": "facebook-ui",
+  "name": "portfolio2",
   "version": "0.1.0",
-  "homepage": "https://Satishrshinde.github.io/portfolio",
   "private": true,
   "dependencies": {
-    "@emailjs/browser": "^3.4.0",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
+    "@emailjs/browser": "^3.11.0",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "cross-env": "^7.0.3",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-anchor-link-smooth-scroll": "^1.0.12",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "predeploy": "npm run build",
     "start": "react-scripts start",
-    "build": "cross-env NODE_ENV=production react-scripts build",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
@@ -39,8 +36,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "gh-pages": "^4.0.0"
   }
 }
