@@ -1,6 +1,8 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./App.css";
-import App from "./App";
-<App />;
-ReactDOM.render(<App />, document.getElementById("root"));
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import "./App.css"
+import App from './App';
+
+const rootElement = document.getElementById('root');
+const root = ReactDOM.createRoot(rootElement);
+root.render(<App />);
